feat: workflow creation mutation now uses toast for errors & file rename
