+++ apps/backoffice-v2/src/pages/Entities/components/CaseCreation/components/CaseCreationForm/hooks/useCaseGenerationForm/index.ts
@@ -1,1 +1,1 @@
-export * from './useCaseGenerationForm';
+export * from './useCaseCreationForm';
