+++ apps/backoffice-v2/src/pages/Entities/components/CaseCreation/components/CaseCreationForm/hooks/useCaseGenerationForm/useCaseCreationForm.ts
